Adds a git alias for quickly merging a branch
Why:

I run the same set of commands whenever I want to merge a branch into
master and it's getting repetitive.

This PR:

Adds a shipit alias that will checkout master, merge the branch I was
on, push and delete remote and local copies of the branch.

Also brings back the commit ammend alias.